Documentation of constraint and elementary constraint.
