+++ AgXUnity/AgXUnity.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Deprecated\Constraint.cs" />
     <Compile Include="Deprecated\ElementaryConstraint.cs" />
     <Compile Include="ElementaryConstraint.cs" />
+    <Compile Include="ElementaryConstraintRowData.cs" />
     <Compile Include="Frame.cs" />
     <Compile Include="Rendering\SegmentSpawner.cs" />
     <Compile Include="Utils\ConstraintUtils.cs" />
